separate out login class, docstrings
