+++ cirro/__init__.py
@@ -1,8 +1,8 @@
 import cirro.file_utils  # noqa
 from cirro.cirro_client import CirroApi
 from cirro.sdk.dataset import DataPortalDataset
+from cirro.sdk.login import DataPortalLogin
 from cirro.sdk.portal import DataPortal
-from cirro.sdk.portal import DataPortalLogin
 from cirro.sdk.process import DataPortalProcess
 from cirro.sdk.project import DataPortalProject
 from cirro.sdk.reference import DataPortalReference
